chore: rename structures to models
