+++ tests/dispense.rs
@@ -2,8 +2,8 @@ use fuel_core::chain_config::{ChainConfig, CoinConfig, StateConfig};
 use fuel_core::service::{Config as NodeConfig, FuelService};
 use fuel_crypto::SecretKey;
 use fuel_faucet::config::Config;
+use fuel_faucet::models::DispenseInfoResponse;
 use fuel_faucet::start_server;
-use fuel_faucet::structures::DispenseInfoResponse;
 use fuel_types::{Address, AssetId};
 use fuels_signers::provider::Provider;
 use fuels_signers::wallet::Wallet;
